Adding the base python code for route collection and adding routes
This is the basic code that I use to collect data on routes and to add routes.  With automatic login set on windows, the computer will restart at 11:59pm each night and start main.py which reads in a text file of routes and sets them 3 seconds apart.  addRoute.py is a command line tool that will automatically add a new route to the text file and create a folder for the route when supplied a name and link.